requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
cn-clip==1.5.1
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
dill==0.3.7
emoji==2.8.0
filelock==3.9.0
fonttools==4.42.1
fsspec==2023.6.0
ftfy==6.1.1
gensim==4.3.2
huggingface-hub==0.16.4
idna==3.4
importlib-resources==6.0.1
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lmdb==1.3.0
MarkupSafe==2.1.2
matplotlib==3.7.2
mpmath==1.2.1
networkx==3.0
nltk==3.8.1
numpy==1.24.1
packaging==23.1
pandas==2.0.3
Pillow==9.3.0
psutil==5.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.28.1
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.10.1
six==1.16.0
smart-open==6.3.0
sympy==1.11.1
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
timm==0.9.7
tokenizers==0.13.3
torch==2.0.0
torch-geometric==2.3.1
torchaudio==2.0.1+cu118
torchdata==0.6.1
torchtext==0.15.2
torchvision==0.15.1+cu118
tqdm==4.66.1
transformers==4.32.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
wcwidth==0.2.6
zipp==3.16.2
